split up cli tests into one file per command prefix
